Independent raw file that contains code
This Python project, titled "Translator," 
accepts text in any language and translates it into the user's desired language. 
It includes features to copy the translated text and paste the text to be translated. 
Additionally, it has an inbuilt play button that speaks the translated text. 
The translator utilizes Googletrans, minimizing the chance of errors.